Refactor MarkPaidTarget class to be a nested class
