+++ src/test/java/seedu/address/logic/parser/MarkPaidCommandParserTest.java
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.MarkPaidCommand;
-import seedu.address.logic.commands.MarkPaidTarget;
 import seedu.address.model.person.MonthPaid;
 
 public class MarkPaidCommandParserTest {
@@ -41,7 +40,7 @@ public class MarkPaidCommandParserTest {
         Index targetIndex = INDEX_SECOND_PERSON;
         String userInput = String.valueOf(targetIndex.getOneBased());
         // empty field
-        MarkPaidCommand expectedCommand = new MarkPaidCommand(MarkPaidTarget.fromIndex(targetIndex),
+        MarkPaidCommand expectedCommand = new MarkPaidCommand(MarkPaidCommand.MarkPaidTarget.fromIndex(targetIndex),
                 Collections.emptySet());
         assertParseSuccess(parser, userInput, expectedCommand);
     }
@@ -64,7 +63,7 @@ public class MarkPaidCommandParserTest {
         Index targetIndex = INDEX_SECOND_PERSON;
         String userInput = String.valueOf(targetIndex.getOneBased()) + " " + VALID_MONTHPAID1_WITH_PREFIX;
         // empty field
-        MarkPaidCommand expectedCommand = new MarkPaidCommand(MarkPaidTarget.fromIndex(targetIndex),
+        MarkPaidCommand expectedCommand = new MarkPaidCommand(MarkPaidCommand.MarkPaidTarget.fromIndex(targetIndex),
                 Set.of(new MonthPaid(VALID_MONTHPAID1)));
 
         assertParseSuccess(parser, userInput, expectedCommand);
@@ -77,7 +76,7 @@ public class MarkPaidCommandParserTest {
                 + " " + VALID_MONTHPAID2_WITH_PREFIX;
         // empty field
         MarkPaidCommand expectedCommand = new MarkPaidCommand(
-                MarkPaidTarget.fromIndex(targetIndex), Set.of(new MonthPaid(VALID_MONTHPAID1),
+                MarkPaidCommand.MarkPaidTarget.fromIndex(targetIndex), Set.of(new MonthPaid(VALID_MONTHPAID1),
                     new MonthPaid(VALID_MONTHPAID2)));
 
         assertParseSuccess(parser, userInput, expectedCommand);
